feat: move helper functions to utils
