+++ pyproject.toml
@@ -39,12 +39,12 @@ treelib = ">=1.7.0"
 setuptools = ">=63.2.0"
 structlog = "^24.1.0"
 dill = "^0.3.8"
-ipykernel = "^6.29.3"
-nbformat = "^5.10.3"
 
 [tool.poetry.group.dev.dependencies]
 pre-commit = "^3.6.0"
 pytest-cov = "^4.1.0"
+ipykernel = "^6.29.3"
+nbformat = "^5.10.3"
 
 [tool.poetry.group.dev]
 optional = true
